+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Change behaviour on API parse errors to write a diagnostic object to the working directory
 * Add `gar_debug_parsing` to help debug API parsing errors.
+* Fix batching that errored if body was exactly the same 
 
 # googleAuthR v0.7.0