+++ NEWS.md
@@ -10,6 +10,8 @@
 * Fully remove `TRAVIS_AUTH` 
 * Remove some message spam from startup
 * Fix mocks failing with `httptest` if used with `with_mock_api()` vs `with_mock_API()`
+* Add `option(googleAuthR.redirect)` to help publish Shiny apps to some platforms (#136)
+
 
 # googleAuthR v0.7.0