+++ NEWS.md
@@ -1,4 +1,4 @@
-# googleAuthR v0.7.0.9000
+# googleAuthR v0.8.0
 
 * Change behaviour on API parse errors to write a diagnostic object to the working directory
 * Add `gar_debug_parsing` to help debug API parsing errors.